pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mindmemory-client"
version = "0.1.0"
description = "Client library: PNMS + MindMemory Web API; includes mmem CLI."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
  "httpx>=0.27",
  "cryptography>=41",
  "pydantic>=2",
  "pydantic-settings>=2",
  "python-dotenv>=1.0",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = ["pytest>=8"]

[project.scripts]
mmem = "mmem_cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
